- image container show loading spinner and notify child if it's drawing - image labeling does not allow change image while container is drawing

+++ ui/src/app/core-ui/image/image-labeling/image-labeling.component.ts
@@ -47,12 +47,14 @@ export class ImageLabelingComponent implements OnInit, AfterViewInit, OnChanges 
   // images
   public imagesSrcs = [];
   public imagesIndex: number;
-  imageDescription: String;
+  imageDescription: string;
 
   public cocoFiles: CocoFormat[] = [];
 
   public isHoverComponent;
   public brushSize: number;
+
+  public isDrawing: boolean = false;
 
   @ViewChild(ImageContainerComponent) imageView: ImageContainerComponent;
 
@@ -286,6 +288,10 @@ export class ImageLabelingComponent implements OnInit, AfterViewInit, OnChanges 
     }
   }
 
+  handleIsDrawing(bool: boolean) {
+    this.isDrawing = bool;
+  }
+
   /* sp-image-labels handler */
   handleLabelChange(label: {category, label}) {
     this.selectedLabel = label;
@@ -293,39 +299,49 @@ export class ImageLabelingComponent implements OnInit, AfterViewInit, OnChanges 
 
   /* sp-image-bar */
   handleImageIndexChange(index) {
-    this.save();
-    this.imagesIndex = index;
+    if (!this.isDrawing) {
+      this.save();
+      this.imagesIndex = index;
+    }
   }
   handleImagePageUp(e) {
-    this.save();
-    this.pageIndex += 1;
-    this.setImagesIndexToLast = true;
-    this.loadData();
+    if (!this.isDrawing) {
+      this.save();
+      this.pageIndex += 1;
+      this.setImagesIndexToLast = true;
+      this.loadData();
+    }
   }
 
   handleImagePageDown(e) {
-    this.save();
-    if (this.pageIndex - 1 >= 0) {
-      this.pageIndex -= 1;
-      this.setImagesIndexToFirst = true;
-      this.loadData();
+    if (!this.isDrawing) {
+      this.save();
+      if (this.pageIndex - 1 >= 0) {
+        this.pageIndex -= 1;
+        this.setImagesIndexToFirst = true;
+        this.loadData();
+      }
     }
   }
 
   /* sp-image-annotations handlers */
   handleChangeAnnotationLabel(change: [Annotation, string, string]) {
-    const coco = this.cocoFiles[this.imagesIndex];
-    const categoryId = this.cocoFormatService.getLabelId(coco, change[1], change[2]);
-    change[0].category_id = categoryId;
-    change[0].category_name = change[2];
-    this.imageView.redrawAll();
+    if (!this.isDrawing) {
+      const coco = this.cocoFiles[this.imagesIndex];
+      const categoryId = this.cocoFormatService.getLabelId(coco, change[1], change[2]);
+      change[0].category_id = categoryId;
+      change[0].category_name = change[2];
+      this.imageView.redrawAll();
+    }
   }
 
   handleDeleteAnnotation(annotation) {
-    if (annotation !== undefined) {
-      const coco = this.cocoFiles[this.imagesIndex];
-      this.cocoFormatService.removeAnnotation(coco, annotation.id);
-      this.imageView.redrawAll();
+    if (!this.isDrawing) {
+      if (annotation !== undefined) {
+        const coco = this.cocoFiles[this.imagesIndex];
+        this.cocoFormatService.removeAnnotation(coco, annotation.id);
+        this.imageView.redrawAll();
+      }
     }
   }
 
@@ -342,7 +358,6 @@ export class ImageLabelingComponent implements OnInit, AfterViewInit, OnChanges 
   }
 
   save() {
-    // TODO
     const coco = this.cocoFiles[this.imagesIndex];
     if (coco !== undefined) {
       const imageSrcSplitted = this.imagesSrcs[this.imagesIndex].split('/');
@@ -359,5 +374,6 @@ export class ImageLabelingComponent implements OnInit, AfterViewInit, OnChanges 
       horizontalPosition: 'right'
     });
   }
+
 
 }
